Fix issue where ratio was always returning "1"

+++ Components/TemplateHelpers/Images/Ratio.cs
@@ -10,7 +10,7 @@ namespace Satrabel.OpenContent.Components.TemplateHelpers
         public int Height { get; private set; }
         public float AsFloat
         {
-            get { return Width / Height; }
+            get { return (float)Width / (float)Height); }
         }
 
         public Ratio(string ratioString)
